pyproject.toml
[project]
name = "open-eyes"
version = "0.1.0"
description = "An telegram bot to assist with OSINT tasks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9,<4",
    "yarl>=1.9,<2",
    "aiogram>=3.4,<4",
    "python-dotenv>=1.0,<2",
    "geopy>=2.4,<3",
    "ddgs>=9,<10",
    "atproto>=0.0.65,<1",
    "feedparser>=6.0,<7",
    "python-dateutil>=2.8,<3",
    "numpy>=2.0,<3",
    "pydantic-ai>=1.57.0",
    "trafilatura>=2.0,<3",
    "playwright>=1.40,<2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: isolated unit tests with mocked dependencies",
    "integration: full component interaction tests",
]
